+++ app/Controllers/MapController.php
@@ -103,6 +103,7 @@ class MapController extends ViewController
 			'username' => $this->session->get('username'),
 			'auth_provider' => $this->session->get('auth_provider'),
 			'current_route' => $this->request->getPathInfo(),
+			'rspamd_stats' => $this->getRspamdStat(),
 		]));
 	}
 
@@ -198,6 +199,7 @@ class MapController extends ViewController
 			'username' => $this->session->get('username'),
 			'auth_provider' => $this->session->get('auth_provider'),
 			'current_route' => $this->request->getPathInfo(),
+			'rspamd_stats' => $this->getRspamdStat(),
 		]));
 	}
 
@@ -297,6 +299,7 @@ class MapController extends ViewController
 			'username' => $this->session->get('username'),
 			'auth_provider' => $this->session->get('auth_provider'),
 			'current_route' => $this->request->getPathInfo(),
+			'rspamd_stats' => $this->getRspamdStat(),
 		]));
 	}
 
@@ -433,6 +436,7 @@ class MapController extends ViewController
 			'username' => $this->session->get('username'),
 			'auth_provider' => $this->session->get('auth_provider'),
 			'current_route' => $this->request->getPathInfo(),
+			'rspamd_stats' => $this->getRspamdStat(),
 		]));
 	}