+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master Dev Branch
 
+### 2025-07-30
+- Added Rspamd statistics for all servers
+
 ## v1.6.0 - Released: 2025-07-29
 
 ### 2025-07-29